fuzz: fuzz ValidatorSet.FromBytes and ToBytes
* Fuzz FromBytes <--> ToBytes

We perhaps need to discuss what to do for
panics in the libraries as currently there
are too many noisy panics from any sensitive
change in the input data slice, since it is
randomized data anyways. It'd be nice
to instead return an error and only panic
when something is critical.

For now, am catching and skipping decoding
errors since randomized data is passed in by
the fuzzer.